backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gitpython==3.1.43
pyyaml==6.0.2
jinja2==3.1.4
